Basic Manhwa Web Scraper
This is a intro web scraper I did to get a better understanding of how they work and different ways to apply the information. I thought it would be an interesting idea to look at how gui would be implemented but I moved to a class project instead so that is why it is half finished.